@@ -12,12 +12,15 @@
     ],
     "config": {
         "platform": {
-            "php": "7.3"
+            "php": "7.4"
         },
         "preferred-install": {
           "pluswerk/grumphp-bom-task": "source",
           "pluswerk/grumphp-xliff-task": "source",
           "*": "dist"
+        },
+        "allow-plugins": {
+            "phpro/grumphp": true
         }
     },
     "support": {
@@ -29,20 +32,20 @@
         }
     },
     "require": {
-        "php": ">=7.3",
+        "php": "~7.4 || ~8.0 || ~8.1",
         "composer-plugin-api": "^1.0.0 || ^2.0.0",
-        "phpro/grumphp": "^1.3.1",
+        "phpro/grumphp": "^1.5.1",
         "pluswerk/grumphp-bom-task": "^7.0.0",
         "pluswerk/grumphp-xliff-task": "^5.0.0",
-        "squizlabs/php_codesniffer": "^3.5.0",
-        "enlightn/security-checker": "^1.3.0",
-        "php-parallel-lint/php-parallel-lint": "^1.0.0",
-        "phpstan/phpstan": "^0.12.87"
+        "squizlabs/php_codesniffer": "^3.6.2",
+        "enlightn/security-checker": "^1.10.0",
+        "php-parallel-lint/php-parallel-lint": "^1.3.2",
+        "phpstan/phpstan": "^1.4.7"
     },
     "extra": {
         "class": "PLUS\\Composer\\Plugin"
     },
     "require-dev": {
-        "composer/composer": "^1.10.22 || ^2.0.13"
+        "composer/composer": "^1.10.25 || ^2.2.7"
     }
 }
